+++ punit-junit5/src/main/java/org/javai/punit/experiment/explore/ExploreStrategy.java
@@ -83,8 +83,7 @@ public class ExploreStrategy implements ExperimentModeStrategy {
                 annotation.samplesPerConfig(),
                 annotation.timeBudgetMs(),
                 annotation.tokenBudget(),
-                annotation.experimentId(),
-                annotation.expiresInDays()
+                annotation.experimentId()
         );
     }